separate the face dection code
