+++ ImageLab/Info.plist
@@ -25,7 +25,7 @@
 	<key>UILaunchStoryboardName</key>
 	<string>LaunchScreen</string>
 	<key>UIMainStoryboardFile</key>
-	<string>ModuleA</string>
+	<string>Main</string>
 	<key>NSCameraUsageDescription</key>
 	<string>Camera needed for live updates</string>
 	<key>UIRequiredDeviceCapabilities</key>
