+++ ImageLab/ModuleAViewController.swift
@@ -13,11 +13,8 @@ import MetalKit
 class MouduleAViewController: UIViewController   {
 
     //MARK: Class Properties
-    var filters : [CIFilter]! = nil
     var videoManager:VisionAnalgesic! = nil
-    let pinchFilterIndex = 2
-    var detector:CIDetector! = nil
-    let bridge = OpenCVBridge()
+    var faceDetector: FaceDetectionModel! = nil
     
     //MARK: Outlets in view
     @IBOutlet weak var cameraView: MTKView!
@@ -25,93 +22,12 @@ class MouduleAViewController: UIViewController   {
     //MARK: ViewController Hierarchy
     override func viewDidLoad() {
         super.viewDidLoad()
-        
         self.view.backgroundColor = nil
-        
-        // setup the OpenCV bridge nose detector, from file
-        //.bridge.loadHaarCascade(withFilename: "nose")
-        
         self.videoManager = VisionAnalgesic(view: self.cameraView)
         self.videoManager.setCameraPosition(position: AVCaptureDevice.Position.front)
-        
-        // create dictionary for face detection
-        // HINT: you need to manipulate these properties for better face detection efficiency
-        let optsDetector = [CIDetectorAccuracy:CIDetectorAccuracyHigh,
-                      CIDetectorNumberOfAngles:11,
-                      CIDetectorTracking:false] as [String : Any]
-        
-        // setup a face detector in swift
-        self.detector = CIDetector(ofType: CIDetectorTypeFace,
-                                  context: self.videoManager.getCIContext(), // perform on the GPU is possible
-            options: (optsDetector as [String : AnyObject]))
-        
-        self.videoManager.setProcessingBlock(newProcessBlock: self.processImageSwift)
-        
-        if !videoManager.isRunning{
-            videoManager.start()
-        }
-    
+        self.faceDetector = FaceDetectionModel(videoManager: self.videoManager)
     }
     
-    //MARK: Process image output
-    func processImageSwift(inputImage:CIImage) -> CIImage{
-        
-        // detect faces
-        let f = getFaces(img: inputImage)
-        
-        // if no faces, just return original image
-        if f.count == 0 { return inputImage }
-        
-        var retImage = inputImage
-        
-        //-------------------Example 1----------------------------------
-        // if you just want to process on separate queue use this code
-        // this is a NON BLOCKING CALL, but any changes to the image in OpenCV cannot be displayed real time
-        /*
-        dispatch_async(dispatch_get_global_queue(DISPATCH_QUEUE_PRIORITY_HIGH, 0)) { () -> Void in
-            self.bridge.setImage(retImage, withBounds: retImage.extent, andContext: self.videoManager.getCIContext())
-            self.bridge.processImage()
-        }
-         */
-        
-        //-------------------Example 2----------------------------------
-        // use this code if you are using OpenCV and want to overwrite the displayed image via OpenCV
-        // this is a BLOCKING CALL
-        /*
-        // FOR FLIPPED ASSIGNMENT, YOU MAY BE INTERESTED IN THIS EXAMPLE
-        
-        self.bridge.setImage(retImage, withBounds: retImage.extent, andContext: self.videoManager.getCIContext())
-        self.bridge.processImage()
-        retImage = self.bridge.getImage()
-         */
-        
-        //-------------------Example 3----------------------------------
-        //You can also send in the bounds of the face to ONLY process the face in OpenCV
-        // or any bounds to only process a certain bounding region in OpenCV
-        
-        self.bridge.setImage(retImage,
-                             withBounds: inputImage.extent, // the first face bounds
-                             andContext: self.videoManager.getCIContext())
-        
-        self.bridge.processFaces(f)
-        retImage = self.bridge.getImageComposite() // get back opencv processed part of the image (overlayed on original)
-        
-        return retImage
-    }
-    
-    //MARK: Setup Face Detection
-    
-    func getFaces(img:CIImage) -> [CIFaceFeature]{
-        // this ungodly mess makes sure the image is the correct orientation
-        let optsFace = [CIDetectorImageOrientation:self.videoManager.ciOrientation,
-                                   CIDetectorSmile:true,
-                                CIDetectorEyeBlink:true] as [String : Any]
-        // get Face Features
-        return self.detector.features(in: img, options: optsFace) as! [CIFaceFeature]
-        
-    }
-    
-        
     //MARK: Convenience Methods for UI Flash and Camera Toggle
     @IBAction func flash(_ sender: AnyObject) {
         if(self.videoManager.toggleFlash()){
@@ -125,7 +41,4 @@ class MouduleAViewController: UIViewController   {
     @IBAction func switchCamera(_ sender: AnyObject) {
         self.videoManager.toggleCameraPosition()
     }
-    
-
-   
 }
